Use protobuf-lite
Set protobuf syntax to "2" in all protobuf files to remove warnings

In order to build with protobuf-lite, protoc-gen-javalite must be in the users PATH.  Follow the instructions at https://github.com/google/protobuf/blob/master/java/lite.md to download.

The biggest change requiring code changes is the removal of get*Builder() methods from the generated Protos

The other change requiring a code change is com.google.protobuf.TextFormat is not available in protobuf-lite.  This class was used in one method that is currently unused, so I removed the method.  The method was most likely used for debugging as it just converted a Wallet to formatted text.